changing meteor's authentication URL to the one that supports login with GitHub (#3)
Change the auth URL from www.meteor.com to account.meteor.com.

www.meteor.com fails when someone tries to log in using their GitHub account. This is because the authentication app created for Meteor on GitHub is set up to redirect to accounts.meteor.com when the authentication is completed.